Reworked DPA ConfigurationMap to use Optional.
After showing the configuration+run harness in DPA (which was used as-was here) to ABR, we agreed
that it made sense to eventually use this approach in other projects, but TRA found it a bit
premature to do it now.  After reading
https://dzone.com/articles/requirements-for-a-configuration-library TRA found that the
"Optional" approach could make things much cleaner and made a trimmed-down version
for this project.