Work with Docker CE 17.09/docker-compose 1.16.1/Nexus mirror on 172.18.95.105.
Upgraded my PC to Ubuntu 17.10 and found that the new stuff I learned about
on DockerCon 2017 EU did not work with the docker I had, so everything had
to be upgraded to a newer version than the one coming with Ubuntu.  As
maven in docker forgets everything I got a Nexus3 running locally and found how
to make the dockerized maven use it.  Now the dockers can launch, and can see
each other but there is not yet any material the ticket checker can validate.